pyproject.toml
[tool.poetry]
name = "biodac-speeding-up-chemistry-simulations"
version = "0.1.0"
description = "Doing chemistry simulations using quantum computing"
authors = ["Benjamin Burghouwt & Bas van der Laan <benjamin.burghouwt@hva.nl>"]
license = "CC0 1.0 DEED"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.3"
qiskit = "1.1.0"
jupyter = "1.0.0"
qiskit-nature = "0.7.2"
pyscf = "2.6.0"
matplotlib = "3.9.0"
tqdm = "4.66.4"
julia = "0.6.2"
openfermion = "1.6.1"
openfermionpyscf = "0.5"
quantuminspire = "2.3.0"
python-dotenv = "1.0.1"
qiskit-ibm-runtime = "0.24.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
